+++ pkg/utils/json.go
@@ -1,4 +1,4 @@
-package crowerutils
+package utils
 
 import (
 	"bytes"
