+++ pkg/utils/url.go
@@ -1,4 +1,4 @@
-package crowerutils
+package utils
 
 import "net/url"
